backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
anthropic==0.40.0
openai==1.51.0
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12
